implement core gameplay
players can make moves on the board and see opponent's moves.
time control and cake rule not yet implemented
began implementing check for win condition. the win condition is
checked on the client side. the proof that win condition is satisfied
should then be sent to the server and checked there.