handle server rejecting a move
new variable out_of_sync added in state of react Game component.
when a move is rejected or something else goes wrong, out_of_sync
can be set to true. then it is indicated to the user that they
should refresh the page.

most APIs have been updated to return a field 'accepted' that
indicates whether the submitted move or other request has been
accepted or not